destructuring: Fix uninitialized access
The “label_lvalue” stack variable isn't initialized when the
destructuring statement also declares the JS variables being assigned
to.

The uninitialized value is passed to “put_lvalue()”, which doesn't
actually make use of it in this particular case. However, memory sanity
checker tools may catch this error and abort execution.

Kudos to @c-smile for the discovery.